requirements.txt
flask
streamlit
requests
python-dotenv
pillow
pandas
